For 43269: Fix project menu duplicated items on invoking 'Reload and Restart'
The fix basically refactors the project menu into a stand alone class.The newly created class handles creation, update and clearing of the project menu.
